requirements.txt
fastapi
pydantic
httpx
loguru
python-dotenv
uvicorn
pydantic-settings
requests
redis
openai
mistral_common
transformers
torch